Build fix for the merge I messed up
git-svn-id: http://svn.ad-group.adh/svn/native_client/native/ffmpeg/trunk@5368 4601b0d2-a6ba-4ff0-a3bb-59d57de6f651

+++ libavcodec/vorbis_dec.c
@@ -1123,8 +1123,8 @@ static int vorbis_floor0_decode(vorbis_context *vc,
     return 0;
 }
 
-static uint_fast8_t vorbis_floor1_decode(vorbis_context *vc,
-                                         vorbis_floor_data *vfu, float *vec)
+static int vorbis_floor1_decode(vorbis_context *vc,
+                                vorbis_floor_data *vfu, float *vec)
 {
     vorbis_floor1 *vf = &vfu->t1;
     GetBitContext *gb = &vc->gb;
